fix: resolve -Weverything -Werror build errors across crate/image/converter
Make native (build) and wasm (web/build_ninja) compile cleanly under
clang -Weverything -Werror:

- stage-converter: initialize WorkItem::depth on child push (was
  -Wmissing-field-initializers); the DFS nesting-depth gate now tracks
  depth correctly instead of always seeing 0 for children.
- usdc-reader-impl: declare the PortalLight and DomeLight_1 explicit
  ReconstructPrim specializations so callers no longer trip
  -Wundefined-func-template (defs live in prim-reconstruct-lightprim.cc).
- crate-writer: use math::is_close(a, b, 0) bit-exact compare for the
  float int-representability checks instead of == / != (-Wfloat-equal),
  matching crate-writer-inline.cc.
- image-writer: rename shadowed EXR `out` vector to `exr_out` (-Wshadow)
  and pragma-wrap the external EXR_OK() macro's C-style cast.
- image-loader: route EXR_OK() through a static_cast-based ExrOk() helper
  (the macro's old-style cast fires at each expansion site).
- imageproc/simd: make int->size_t index arithmetic explicit
  (-Wsign-conversion).
- imageio/png-stream: convert C-style casts to static_cast, make
  int->size_t conversions explicit, and pragma-wrap the external
  miniz.h / stb_image_resize2.h includes.

No behavior change. ctest 21/21 and usdcat roundtrip 217/217 (+1 XFAIL) pass.

+++ src/imageproc/simd.cc
@@ -84,10 +84,11 @@ void PackChannels8(uint8_t *out, size_t n_pixels, int out_channels,
     if (s.in) {
       const uint8_t *in = s.in + s.channel;
       const int istr = s.in_stride;
-      for (size_t x = 0; x < n_pixels; ++x) out[x * oc + c] = in[x * istr];
+      for (size_t x = 0; x < n_pixels; ++x)
+        out[x * size_t(oc) + size_t(c)] = in[x * size_t(istr)];
     } else {
       const uint8_t v = s.constant;
-      for (size_t x = 0; x < n_pixels; ++x) out[x * oc + c] = v;
+      for (size_t x = 0; x < n_pixels; ++x) out[x * size_t(oc) + size_t(c)] = v;
     }
   }
 }
